linux build

+++ src/32_geography/MapSourceUpdateAction.cpp
@@ -68,7 +68,7 @@ namespace synthese
 			}
 			if(_coordinatesSystem)
 			{
-				map.insert(PARAMETER_SRID, _coordinatesSystem->getSRID());
+				map.insert(PARAMETER_SRID, static_cast<int>(_coordinatesSystem->getSRID()));
 			}
 			if(_type)
 			{
